Move PackageBuilder converters to core package
To improve future refactoring and decoupling from UI frameworks the
previously internal type description classes are moved away from the
dialog and into the core platform agnostic package.
