+++ Bonsai.NuGet/Bonsai.NuGet.csproj
@@ -4,11 +4,8 @@
     <GenerateDocumentationFile>false</GenerateDocumentationFile>
     <TargetFrameworks>net472;net8.0</TargetFrameworks>
   </PropertyGroup>
-  <ItemGroup Condition="'$(TargetFrameworkIdentifier)' == '.NETFramework'">
-    <PackageReference Include="Rx-Main" Version="2.2.5" />
-  </ItemGroup>
-  <ItemGroup Condition="'$(TargetFrameworkIdentifier)' != '.NETFramework'">
-    <PackageReference Include="System.Reactive" Version="6.0.1" />
+  <ItemGroup>
+    <ProjectReference Include="..\Bonsai.Core\Bonsai.Core.csproj" />
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="NuGet.Protocol" Version="6.6.1" />
